+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
- - Turn off `-Wmisleading-indentation` only when `__GNUC__ >= 6`.
+ - Suppress `-Wmisleading-indentation` only when `__GNUC__ >= 6`.
+ - Suppress `-Wreturn-type` for Clang too.
  - Make it work on C++11.
 
 ## [0.3.0] - 2021-02-26